Export metacache & shutterrc directly from main entry point
