--- packages/core/metacache.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/metacache/index'
