--- packages/core/shutterrc.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/shutterrc'
