--- packages/core/shutterrc.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/shutterrc')
